[BUGFIX] fix update script for TYPO3 8
